add changed pixels

+++ src/ms_pro/utils/utils.h
@@ -7,8 +7,8 @@ enum class figure_t
     RECTANGLE = 0,
 };
 
-float const kMinSize = 1;
-float const kMaxSize = 300;
+int const kMinSize = 1;
+int const kMaxSize = 300;
 //figure_t kFirstFigure = figure_t::CIRCLE;
 //QColor kColBack = "#F1CFED";
 //QColor kColPixel = Qt::blue;
